+++ src/Federation/BatchedEntityResolver.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Nuwave\Lighthouse\Federation;
+
+interface BatchedEntityResolver
+{
+    /**
+     * @param array<int, array<string, mixed>> $representations
+     *
+     * @return iterable<mixed> must be the same count as $representations
+     */
+    public function __invoke(array $representations): iterable;
+}